Add input tape verification
+ Fix small typo in the declaration of dynamic predicates
+ Add the tape verification to the step/0 predicate

The predicates verify_tapes/2 and verify_tape/1 were added.
They will present a error message and then fail if any of the following
occur:

  - The set of (partial) tapes is not a list
  - The set of the left parts and the set of the right parts don't have
    equal lengths
  - If any of the left or right parts is not a list
  - If any symbol is actually a (Prolog) variable

These checks were only included in the stepping exection, so that the running
one can have the best possible performance, by trusting that the user
has submitted correct inputs.

On branch enhancements

Changes to be committed:
	modified:   lib/turing_base.pl
	modified:   lib/turing_step.pl