Modify tapes_initial_state/2 to reverse left tapes
During the simulation, the left tape is in reversed order. This means
that the most central element of the left tape is actually the first,
and so, when printed, it appears to be the leftmost, and viceversa for
the last element.

With this change, instead of having to write:

    [$, 0, 1, 0]

where `$` is the most central element, users must write:

    [0, 1, 0, $]

This inversion of the left is also taken in account when showing the
left tape (shows the "natural" order).

On branch enhancements
Changes to be committed:
	modified:   lib/turing_base.pl