Fix repeated state showing up
The bug: Sometimes, the predicate show_machine_state/8 appeared twice,
and only asked at the second repetition for input (to the next step).

The cause: After show_machine_state/8, the predicate
tapes_movimentations/5 was tried. That predicate assumes that the right
part of the tape has, at least, 1 element -- the head symbol.

However, there was one case where that was not happening.

To get the symbols that are "under" the head, we use a predicate called
tape_head/3. It is composed only by 2 facts, and they had the following
order:

    tape_head([], null, []).
    tape_head([H | T], H, T).

As you can see, the first fact is to safeguard the case when there were
no more symbols on the right part of the tape -- in that case, a Turing
Machine reads the empty symbol. In this implementation, that void is
represented by the atom `null`.

The problem arised when we wanted the first argument, giving the 2nd and
3rd arguments -- this operation is how the symbols are overwritten,
acording to the rules. If the 2nd argument was `null`, and the rest of
the right part of the tape is empty (3rd argument), both facts may
apply. This means that, and because of the ordering of the facts,
instead of the result being `[null]`, the result was `[]`.

This broke the assumption that a right tape has always one symbol, and
forced a retry. That retry, occurring in tapes_movimentations/5, ended
when we retried to prove tapes_heads/3, and in that case the result was
`[null]`, and tapes_movimentations succeeded. But that made
show_machine_state/8 execute twice.

The fix: change the ordering of the tape_head/3 facts, so that the
result of joining the `null` atom with an empty list is `[null]`. This
fix solves all the mentioned problems.

Changes to be committed:
	modified:   lib/turing_base.pl